refactor(Post): create PostSkeleton and combine Components
