+++ src/components/Post/Post.tsx
@@ -1,0 +1,24 @@
+import Content from './PostContent';
+import ContentTitle from './PostTitle';
+import type { FrontMatter, Post as PostModel } from '~/models/Post';
+
+interface Props {
+  post: PostModel;
+}
+
+const Post = ({ post }: Props) => {
+  const frontMatter: FrontMatter = {
+    title: post.title,
+    tags: post.tags,
+    description: post.description,
+    createdTime: post.createdTime,
+  };
+  return (
+    <>
+      <ContentTitle postMetaData={{ ...frontMatter }} />
+      <Content content={post.content} />
+    </>
+  );
+};
+
+export default Post;
