+++ src/pages/test/index.tsx
@@ -1,10 +1,10 @@
-import ContentSkeleton from '~/components/Post/ContentSkeleton';
+import PostSkeleton from '~/components/Post/PostSkeleton';
 import ContentLayout from '~/components/layouts/ContentLayout';
 
 const Page = () => {
   return (
     <ContentLayout>
-      <ContentSkeleton />
+      <PostSkeleton />
     </ContentLayout>
   );
 };
